add icon_name
In X11, a window and its icon (as displayed in window managers that
show an icon on the root window when a window is iconified; and perhaps
in window managers that iconify in a more "modern" way) can have
distinct names. This addition works just like the `name` file under
window directories, but pertains to the icon name instead.